revert Domoticz pull #4034
Sending local times from a client is a bad idea, particularly when
they have been formatted to look like UTC times. In communication
with the server a single time reference should be used to avoid
translation conflicts and as the client may (temporarily) be in a
different time zone the logical choice is UTC, which is why this
the default behaviour of the javascript toISOString() function.

In case communication with specific target systems, e.g. `Evohome
Serial` requires local times to be sent, then this should be
handled by the hardware component using the server's time settings.