AP Systems ECU-R: need to notify eventsystem for scripts to work
New firmware update for the ECU-R appears to cause the device to
refuse local network querying more frequently than before. This
situation requires us to power cycle the ECU-R and to automate
this we need the eventsystem to be aware of status updates, which
prior to this was not the case.